Aufteilung der Radioklasse aus asksinpp übernommen
